GH-1062 - Stratagem duration models should be bigint (#1068)
* GH-1062 - Stratagem duratiion models should be bigint

Since milleseconds will be used for duration and it's possible that a
number of years could be specified, the fields should be of the
BigInteger type instead of Integer. This patch will ensure that:
1. The duration fields on the models use BigInteger
2. The API will handle hydrating and dehydrating data from String to
BigInteger such that data is stored as a u64 in the database and
retrieved as a long instead of a string (Tastypie does not support
    BigInteger in resources so it must be converted to a long)
3. Validation is added to ensure that no value > MAX_SAFE_INTEGER (9007199254740991)
  can be used for either duration.


* - Some cleanups


* - Updates


* - Upates


* - Refactor validation class to use composition. This will clean up the
validation function and reduce complexity.


* - Some small cleanups


* - remove unused import

+++ chroma_core/models/stratagem.py
@@ -42,10 +42,10 @@ class StratagemConfiguration(StatefulObject):
         help_text="The filesystem id associated with the stratagem configuration", null=False
     )
     interval = models.IntegerField(help_text="Interval value in seconds between each stratagem execution", null=False)
-    report_duration = models.IntegerField(
+    report_duration = models.BigIntegerField(
         help_text="Interval value in seconds between stratagem report execution", null=True
     )
-    purge_duration = models.IntegerField(help_text="Interval value in seconds between stratagem purges", null=True)
+    purge_duration = models.BigIntegerField(help_text="Interval value in seconds between stratagem purges", null=True)
 
     states = ["unconfigured", "configured"]
     initial_state = "unconfigured"
@@ -200,8 +200,8 @@ class StreamFidlistStep(Step):
 class RunStratagemJob(Job):
     mdt_id = models.IntegerField()
     uuid = models.CharField(max_length=64, null=False, default="")
-    report_duration = models.IntegerField(null=True)
-    purge_duration = models.IntegerField(null=True)
+    report_duration = models.BigIntegerField(null=True)
+    purge_duration = models.BigIntegerField(null=True)
     fqdn = models.CharField(max_length=255, null=False, default="")
     target_name = models.CharField(max_length=64, null=False, default="")
     filesystem_type = models.CharField(max_length=32, null=False, default="")
@@ -338,8 +338,8 @@ class SendResultsToClientStep(Step):
 
 class SendStratagemResultsToClientJob(Job):
     uuid = models.CharField(max_length=64, null=False, default="")
-    report_duration = models.IntegerField(null=True)
-    purge_duration = models.IntegerField(null=True)
+    report_duration = models.BigIntegerField(null=True)
+    purge_duration = models.BigIntegerField(null=True)
 
     class Meta:
         app_label = "chroma_core"
